frontend: Move TextInputComponent and Checkbox InputComponent to input directory inside components
