+++ frontend/src/components/input/CheckboxInputComponent.vue
